+++ tools/docs/pypi_zerobug/conf.py
@@ -21,7 +21,7 @@
 # -- Project information -----------------------------------------------------
 
 project = 'zerobug'
-copyright = '2019-22, SHS-AV s.r.l.'
+copyright = '2019-23, SHS-AV s.r.l.'
 author = 'Antonio Maria Vigliotti'
 
 # The short X.Y version