+++ wok_code/wok_code/install_python_3_from_source.sh
@@ -1,9 +1,10 @@
 #!/usr/bin/env bash
-[[ ! $1 =~ ^(3.6|3.7|3.8) ]] && echo "$0 3.6|3.7|3.8" && exit 1
+[[ ! $1 =~ ^(3.6|3.7|3.8|3.9) ]] && echo "$0 3.6|3.7|3.8|3.9" && exit 1
 [[ $EUID -ne 0 ]] && echo "This code may be executed just by root user" && exit 1
 [[ -z $(which wget 2>/dev/null) ]] && echo "Please install wget" && exit 1
 pyver="$1"
-[[ $1 == "3.8" ]] && pyver="$1.10"
+[[ $1 == "3.9" ]] && pyver="$1.15"
+[[ $1 == "3.8" ]] && pyver="$1.15"
 [[ $1 == "3.7" ]] && pyver="$1.15"
 [[ $1 == "3.6" ]] && pyver="$1.15"
 echo "If you want to use sqlite you should install some packages before this command"