+++ wok_code/wok_code/scripts/deploy_odoo.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2018-22 SHS-AV s.r.l. (<http://ww.zeroincombenze.it>)
+# Copyright 2018-23 SHS-AV s.r.l. (<http://ww.zeroincombenze.it>)
 #
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 #
@@ -775,7 +775,7 @@ def main(cli_args=None):
     cli_args = cli_args or sys.argv[1:]
     parser = argparse.ArgumentParser(
         description="Manage Odoo repositories",
-        epilog="© 2021-2022 by SHS-AV s.r.l."
+        epilog="© 2021-2023 by SHS-AV s.r.l."
     )
     parser.add_argument("-A", "--update-addons-conf",
                         action="store_true",