+++ wok_code/wok_code/scripts/generate_random_codes.py
@@ -7,7 +7,7 @@ from random import random, randint
 import vatnumber
 
 
-__version__ = "2.0.4"
+__version__ = "2.0.3.1"
 
 
 def gen_vatnumber(opt_args):
@@ -27,7 +27,7 @@ def main(cli_args=None):
     cli_args = cli_args or sys.argv[1:]
     parser = argparse.ArgumentParser(
         description="Generate random VAT number",
-        epilog="© 2021-2022 by SHS-AV s.r.l."
+        epilog="© 2021-2023 by SHS-AV s.r.l."
     )
     parser.add_argument('-i', '--iso', default="IT")
     parser.add_argument('-v', '--verbose', action='count', default=0)